Stop getting rate limited by GitHub Actions when installing protoc
Since the `arduino/protoc` action is using the Github API to download and install the `protoc` binary, we were getting rate limited by Github Actions. This was causing the CI to fail sometimes.

As suggested by the `arduino/protoc` action documentation, we can pass a `GITHUB_TOKEN` to the action to avoid getting rate limited.

Ref: https://github.com/arduino/setup-protoc?tab=readme-ov-file#usage