Reuse the main buffer for encoding submessages
Previously, encoding sub-messages would allocate a new string so that the
parent message could capture the length of the sub-message and encode
that length.  (The encoded sub-message length must appear before the
encoded sub-message itself)

This change forces sub-messages to reuse the same buffer as parent
messages.  We use `String#bytesize` on the message buffer to calculate
the sub-message size, then use `String#bytesplice` to insert enough
bytes to encode the sub-message length