Use `String#append_as_bytes` if available
Final version of: https://github.com/Shopify/protoboeuf/pull/116
Ref: https://bugs.ruby-lang.org/issues/20594

This isn't yet as fast as the prototype because the final method
was merged as `append_as_bytes` with variadic and Integer support
so I haven't implemented a YJIT acceleration yet.

But since we can now consider the interface stable, we can merge.