Optimize `encode_submessage` to avoid shifting for small messages
An optimization idea discussed with Maxime.

Right now we reserve 10 bytes for submessages, then shift the
submessage in the buffer to get rid of the unused ones.

This means we're always shifting the submessage unless it's
gigantic enough to use 10 bytes to encode its length.

If instead we only reserve one byte, it allows us to not shift
anything if the submessage is small enough for its length to
only need one byte. So less work to do for submessages of less
than 127B, as about as much work as before for bigger ones.