Always do the full varint reading when pulling the tag
A message (written by a newer version) may contain a tag number
higher than 15 (even if this older version has less than 15).

It may be possible to reintroduce this optimization
but the "if not found" code will have to determine if it needs
more bytes to finish the tag
before trying to pull the value.